pyproject.toml
[project]
name = "pdx"
version = "0.1.0"
license = { text = "Apache-2.0" }
requires-python = ">=3.13"
dependencies = [
    "click",
    "open-clip-torch",
    "pillow",
    "pillow-heif",
    "qdrant-client",
    "torch",
    "torchvision",
    "tqdm",
    "requests",          # Added for Ollama API
    "reverse_geocoder",  # Added for offline GPS
    "pyyaml"             # Added to handle config.yaml parsing
]

[project.optional-dependencies]
dev = [
    "ipdb",
    "ruff",
    "ty",
]

[project.scripts]
pdx = "pdx.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
